Refactor world generation and controls into modules
